Refactor: archiveDownloader
